# AV2_SBD/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sistema Aeroporto",
    "version": "1.0.0",
    "description": "API para gerenciar voos, passageiros e infraestrutura aeroportuária."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor Local"
    }
  ],
  "paths": {
    "/voos": {
      "get": {
        "summary": "Listar todos os voos",
        "responses": {
          "200": {
            "description": "Lista de todos os voos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Voo" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo voo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Voo" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voo criado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Voo" }
              }
            }
          }
        }
      }
    },
    "/voos/{id}": {
      "get": {
        "summary": "Detalhes de um voo específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do voo",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Voo" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar dados do voo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Voo" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voo atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Voo" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletar um voo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Voo deletado com sucesso"
          }
        }
      }
    },
    "/voos/{id}/status": {
      "patch": {
        "summary": "Alterar o status do voo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status do voo atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Voo" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Voo": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "codigo_voo": { "type": "string" },
          "origem": { "type": "string" },
          "destino": { "type": "string" },
          "horario_partida": { "type": "string", "format": "date-time" },
          "horario_chegada": { "type": "string", "format": "date-time" },
          "status": { "type": "string" }
        },
        "example": {
          "id": 1,
          "codigo_voo": "AB123",
          "origem": "São Paulo",
          "destino": "Rio de Janeiro",
          "horario_partida": "2024-11-26T08:00:00Z",
          "horario_chegada": "2024-11-26T09:30:00Z",
          "status": "ativo"
        }
      }
    }
  }
}